requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
rich==13.9.3
